+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+
 ### Added
 
 ### Changed
+
+- rateLimitミドルウェアのRedisキー名を変更(ディリミタをドットからコロンに変更)
+- update packages
 
 ### Deprecated
 
@@ -18,10 +22,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 
 ## v0.1.0 - 2018-02-04
+
 ### Added
+
 - Cognito認証ミドルウェアを追加。
 
 ## v0.0.0 - 2017-11-28
+
 ### Added
+
 - ベーシック認証ミドルウェアを追加。
 - レート制限ミドルウェアを追加。